server/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
razorpay==1.4.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
dj-database-url==2.1.0
gunicorn==21.2.0
supabase==2.4.5
